장바구니, 유저 메인 페이지, 메뉴 상세 페이지 구현 (#10)
* 메뉴상세피이지

* MenuMapper

* 임시 저장/장바구니

* 장바구니

---------

+++ src/main/java/com/cafein/menu/MenuDTO.java
@@ -9,8 +9,8 @@ public class MenuDTO {
     private Long menuId;
     private String name;
     private String content;
-    private Long price;
-    private String menuPictureUrl; // CLOB → String
+    private int price;
+    private String menuPictureUrl;
     private Date createdDate;
     private Date modifiedDate;
 }
